Major refactoring changes
The main change is in the DataHandler. All received data is now only
stored in a binary file in non volatile memory.
This also means the different modes, like passthrough mode, are not
valid anymore. Future commands have yet to be defined and until then
the old commands will remain as comments in the respective source file.